Good functionality, still a few bugs and design to be done
For some reason, there is a strange behavior when i comment or uncomment the line 75 in the index file. The alert when triggered makes the page go through the whole newgame process after just one click of the "nouvelle partie" btn, but if it is not run, we have to click twice to start the new game.